+++ HISTORY.md
@@ -13,6 +13,7 @@ https://github.com/josdejong/mathjs
 - Renamed unit conversion function and operator `in` to `to`. Operator `in` is
   still available in the expression parser as an alias for `to`. Added unit
   `in`, an abbreviation for `inch`. Thanks Elijah Insua (tmpvar).
+- Added plurals and aliases for units.
 - Implemented an argument `includeEnd` for function `range` (false by default).
 - Ranges in the expression parser now support big numbers.
 - Implemented functions `permutations` and `combinations`.